chore: Renamed docs-app-for-embroider-css-modules-temporary to my-app-with-ember-css-modules
