+++ .changeset/angry-cars-raise.md
@@ -1,7 +1,7 @@
 ---
 "ember-codemod-remove-ember-css-modules": patch
 "embroider-css-modules-temporary": patch
-"docs-app-for-embroider-css-modules-temporary": patch
+"my-app-with-ember-css-modules": patch
 "embroider-css-modules": patch
 "test-app-for-embroider-css-modules": patch
 "my-app": patch
