+++ docs/my-app-with-ember-css-modules/.editorconfig
