+++ docs/my-app-with-ember-css-modules/.prettierrc.js
