+++ docs/my-app-with-ember-css-modules/.stylelintrc.js
