+++ docs/my-app-with-ember-css-modules/.watchmanconfig
