+++ docs/my-app-with-ember-css-modules/README.md
@@ -1,7 +1,7 @@
 [![This project uses GitHub Actions for continuous integration.](https://github.com/ijlee2/embroider-css-modules/actions/workflows/ci.yml/badge.svg)](https://github.com/ijlee2/embroider-css-modules/actions/workflows/ci.yml)
 [![This project is using Percy.io for visual regression testing.](https://percy.io/static/images/percy-badge.svg)](https://percy.io/Isaac/embroider-css-modules)
 
-# docs-app-for-embroider-css-modules-temporary
+# my-app-with-ember-css-modules
 
 1. [What is it?](#what-is-it)
 1. [Local development](#local-development)
@@ -12,7 +12,7 @@
 
 ## What is it?
 
-`docs-app-for-embroider-css-modules-temporary` is an Ember app. We use it to check that `ember-css-modules` and `embroider-css-modules-temporary` can coexist.
+`my-app-with-ember-css-modules` is an Ember app with `ember-css-modules`. We use it to check that `ember-css-modules` and `embroider-css-modules` can coexist.
 
 
 ## Local development
@@ -24,7 +24,7 @@ Before starting the application, build its dependencies so that you can test the
 pnpm build
 
 # Change directory
-cd docs/embroider-css-modules-temporary
+cd docs/my-app-with-ember-css-modules
 ```
 
 Some useful commands:
