+++ docs/my-app-with-ember-css-modules/app/config/environment.d.ts
