+++ docs/my-app-with-ember-css-modules/app/index.html
@@ -9,7 +9,7 @@
     {{content-for "head"}}
 
     <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/docs-app-for-embroider-css-modules-temporary.css">
+    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/my-app-with-ember-css-modules.css">
 
     {{content-for "head-footer"}}
   </head>
@@ -17,7 +17,7 @@
     {{content-for "body"}}
 
     <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/docs-app-for-embroider-css-modules-temporary.js"></script>
+    <script src="{{rootURL}}assets/my-app-with-ember-css-modules.js"></script>
 
     {{content-for "body-footer"}}
   </body>
