+++ docs/my-app-with-ember-css-modules/app/styles/app.css.d.ts
