+++ docs/my-app-with-ember-css-modules/app/styles/index.css.d.ts
