+++ docs/my-app-with-ember-css-modules/app/templates/application.hbs
