+++ docs/my-app-with-ember-css-modules/config/ember-cli-update.json
