+++ docs/my-app-with-ember-css-modules/config/optional-features.json
