+++ docs/my-app-with-ember-css-modules/config/targets.js
