+++ docs/my-app-with-ember-css-modules/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "docs-app-for-embroider-css-modules-temporary",
+  "name": "my-app-with-ember-css-modules",
   "version": "1.0.0",
   "private": true,
-  "description": "docs-app for embroider-css-modules-temporary",
+  "description": "An Ember app with ember-css-modules",
   "keywords": [
     "css",
     "css-modules",
