+++ docs/my-app-with-ember-css-modules/tests/acceptance/index/accessibility-test.ts
