+++ docs/my-app-with-ember-css-modules/tests/acceptance/index/as-user-test.ts
