+++ docs/my-app-with-ember-css-modules/tests/acceptance/index/visual-regression-test.ts
