+++ docs/my-app-with-ember-css-modules/tests/helpers/percy.ts
@@ -10,7 +10,7 @@ type QUnitAssert = {
 };
 
 function getName(qunitAssert: QUnitAssert): string {
-  const packageName = 'docs-app-for-embroider-css-modules-temporary';
+  const packageName = 'my-app-with-ember-css-modules';
   const moduleName = qunitAssert.test.module.name;
   const testName = qunitAssert.test.testName;
