+++ docs/my-app-with-ember-css-modules/tests/index.html
@@ -10,7 +10,7 @@
     {{content-for "test-head"}}
 
     <link rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link rel="stylesheet" href="{{rootURL}}assets/docs-app-for-embroider-css-modules-temporary.css">
+    <link rel="stylesheet" href="{{rootURL}}assets/my-app-with-ember-css-modules.css">
     <link rel="stylesheet" href="{{rootURL}}assets/test-support.css">
 
     {{content-for "head-footer"}}
@@ -30,7 +30,7 @@
     <script src="/testem.js" integrity="" data-embroider-ignore></script>
     <script src="{{rootURL}}assets/vendor.js"></script>
     <script src="{{rootURL}}assets/test-support.js"></script>
-    <script src="{{rootURL}}assets/docs-app-for-embroider-css-modules-temporary.js"></script>
+    <script src="{{rootURL}}assets/my-app-with-ember-css-modules.js"></script>
     <script src="{{rootURL}}assets/tests.js"></script>
 
     {{content-for "body-footer"}}
