+++ docs/my-app-with-ember-css-modules/types/global.d.ts
@@ -1,5 +1,5 @@
 // Types for compiled templates
-declare module 'docs-app-for-embroider-css-modules-temporary/templates/*' {
+declare module 'my-app-with-ember-css-modules/templates/*' {
   import { TemplateFactory } from 'ember-cli-htmlbars';
 
   const tmpl: TemplateFactory;
