+++ docs/my-app-with-ember-css-modules/types/my-app-with-ember-css-modules/index.d.ts
